PB-239: Move email validation function to utils.
