--- src/modules/menu/components/menu/common/utils.js
@@ -1,6 +1,0 @@
-export function isValidEmail(email) {
-    // comes from https://v2.vuejs.org/v2/cookbook/form-validation.html#Using-Custom-Validation
-    const EMAIL_REGEX =
-        /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-    return EMAIL_REGEX.test(email)
-}
