+++ src/utils/utils.js
@@ -217,3 +217,16 @@ export function transformUrlMapToEmbed(url) {
     }
     return urlObj.toString()
 }
+
+/**
+ * Check if the provided string is a valid email address
+ *
+ * @param {string} email Email address to check
+ * @returns {boolean} True if valid, false otherwise
+ */
+export function isValidEmail(email) {
+    // comes from https://v2.vuejs.org/v2/cookbook/form-validation.html#Using-Custom-Validation
+    const EMAIL_REGEX =
+        /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+    return EMAIL_REGEX.test(email)
+}
